requirements.txt
Babel==2.7.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Babel==0.12.2
Flask-SQLAlchemy==2.4.1
Flask-Table==0.5.0
Flask-WTF==0.14.2
gav4==0.1.2
google-api-python-client==1.7.11
google-apputils==0.4.2
google-auth==1.7.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gspread==3.1.0
gunicorn==19.9.0
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numexpr==2.7.0
numpy==1.17.3
oauth2client==4.1.3
oauthlib==3.1.0
psycopg2==2.7.6.1
pyasn1==0.4.7
pyasn1-modules==0.2.7
python-dateutil==2.8.1
python-gflags==3.1.2
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.9
tables==3.6.1
uritemplate==3.0.0
urllib3==1.25.6
Werkzeug==0.16.0
WTForms==2.2.1
